Refactored code to extract bonus product item
