+++ packages/template-retail-react-app/app/components/bonus-product-item/bonus-product-item.jsx
