+++ packages/template-retail-react-app/app/components/bonus-product-modal/index.jsx
@@ -21,7 +21,7 @@ import {
 import {useProducts} from '@salesforce/commerce-sdk-react'
 import {useBonusProductModalContext} from '@salesforce/retail-react-app/app/hooks/use-bonus-product-modal'
 import {FormattedMessage} from 'react-intl'
-import BonusProductItem from 'components/bonus-product-modal/bonus-product-item'
+import BonusProductItem from '@salesforce/retail-react-app/app/components/bonus-product-item/bonus-product-item'
 
 export const BonusProductModal = () => {
     const {isOpen, onClose, data} = useBonusProductModalContext()
